WorkflowSettingsPage: fix role=tabpanel overwriting form role - Also, 'tab' data moved up from WorkflowHeader to PagesEditor.
